% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jumuisha.R
\name{jumuisha}
\alias{jumuisha}
\title{Jumuisha}
\usage{
jumuisha(x, y)
}
\arguments{
\item{x}{First number}

\item{y}{Second number}
}
\value{
A sum of x and y
}
\description{
A function that takes two numbers and add them
}
\examples{
jumuisha(10, 30)
}
