% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squisher.R
\name{squisher}
\alias{squisher}
\title{squisher}
\usage{
squisher(vec)
}
\arguments{
\item{vec}{A vector of characters}
}
\value{
The same vector without leading or trailing spaces
}
\description{
A function that removes trailing and leading spaces from characters
}
\examples{
squisher(c("Shel ", " Shiku", "Nimu "))
}
